[runners] Use argparse for parsing options in run_all_tests
This patch completely refactors the way `run_all_tests` parses the command
line options. As a result, the order of options doesn't matter anymore but
each parameter requires a '-XXX' preceding argument. Concretely, the filter
param requires -r, the timeout requires -t and the new test-type param
requires -T. In addition to that, now -t and -r have effect even when we're
just getting the list of tests with -l.