[self_tests] Fix a bug in se_find()
The function was inconditionally discarding the anything following the
last '_' in the test name. That was OK for special suffixes like "_med"
and "_short", but it wasn't for tests like "kmutex_ord_med", when passed
manually by the user in their shorter form as "kmutex_ord": in that case,
the "_ord" was dropped and the "kmutex_med" test was actually performed
instead. That was a pretty nasty behavior, but fortunately, it affected
*only* manual test runs via the `selftest` command. Test runners always
used the full test name, including the special suffix.