[devshell] Don't attempt to run <selftest>_manual
In the past, the sys_tilck_cmd() syscall expected a full self-test name,
(<name> + {_short, _med, _long, _manual}) when cmd = TILCK_CMD_RUN_SELFTEST.
In order to run at least the manual tests in an easier way, cmd_selftest()
used to try appending "_manual" to the test name in case the syscall failed.
Because now that's not the case anymore, and self-tests are guaranteed
to have a unique suffix-less name, that 2nd attempt is not needed anymore.