[azure] Add trigger block with branches and paths to exclude
It is very convenient to automatically skip CI builds in case of branches
named 'readme' or temp-* or exp-* in order to avoid wasting resources.
Also, the trigger block allows to automatically skip building any branches
that touched just a given set of files (like README.md and doc files).