[interactive] Drop the test regex filter feature
The `run_interactive_test` script is the analogue of `single_test_run`
which is supposed to run just one test. It is not a multiple test runner:
that's the role of `run_all_tests`. As now there's an on-going effort
to make `run_all_tests` being able to run interactive tests as well, it
makes to drop redundant features from this script.