[x86_strings] Fix corner case in memcpy's ASSERT
In order to avoid overlapping buffers memcpy() had the following ASSERT:

	ASSERT( dest < src || (src + n <= dest) )

The problem with that it didn't handle the weird case where dest == src.
Therefore, in this patch the ASSERT has been replaced with:

	ASSERT( dest <= src || (src + n <= dest) )

In addition, this patch adds additional comments in the memcpy() to allow
any reader to immediately understand all the cases involved.

BONUS: the check has been fixed for memcpy32() and memmove() too.