[printk] Don't use READ_U32() with the NO_PREFIX string literal
With commit `c51ba65d` which introduced READ_U32(), in printk() we converted
the "*(u32 *)NO_PREFIX" cast to READ_U32(NO_PREFIX), which is safer against UB
because it uses __builtin_memcpy() instead of potentially performing an
unaligned access. The problem is that now, with GCC 9.x, we get the following
error while compiling the unit tests:

	error: ‘__builtin_memcpy’ reading 4 bytes from a region of size 1
	[-Werror=stringop-overflow=]

And the compiler's right because NO_PREFIX is defined as:

	#ifndef UNIT_TEST_ENVIRONMENT
	   #define NO_PREFIX          "\x01\x01\x20\x20"
	#else
	   #define NO_PREFIX          ""
	#endif

And for the unit tests, its UB to read 4 bytes from an empty string.
This patch fixes the problem by:

	1. Reading a single char `cmd` at fmt[1] instead of a 32-bit word.
	2. Checking that NO_PREFIX[0] is != 0 before reading anything else
	3. Comparing just NO_PREFIX[1] to the current single-char `cmd`.