[start] Map the first 8 MB at KERNEL_BASE_VA in start.S
Tilck is loaded at its physical address +1 MB and assumes the paging is
disabled or is enabled with identity-mapping. Because the C code actually
works only when the kernel is loaded at its *virtual* address, in start.S
we need to setup the paging in a temporary way just to map at least the
kernel itself at its vaddr and jump to its high (virtual) address. The code
for doing that, exists since the beginning, and it used to map the just first
4 MB.

Now, with certain sanitizers (e.g. UBSAN) kernel's code size explodes and
gets bigger than 4 MB. Therefore, we need to map more than 4 MB in start.S.
Because we use big pages, just map the first 8 MB.