[process] Bugfix: map isolated stack blocks in the __kernel_pdir
Previously, process' pdir was used and that was wrong because kernel's stack
is freed *after* process' pdir is destroyed. That has never been a problem
with the regular kernel stack, because it didn't require any mapping. With the
isolated stack (debug feature) instead, we mapped the "stack block" in the
"high vmem" using process' pdir. That was simply wrong: we should have used
the "kernel pdir" instead, which is never destroyed.