[task][process] Rename 'arch_fields' to ti_arch and pi_arch in structs
Previously, both in struct task and in struct process, the opaque
arch-specific data was called 'arch_fields'. It seemed good to use the
same name in both the cases, but in practice it made possible the macro
get_task_arch_fields() to be used instead of get_proc_arch_fields()
and vice-versa.