[process] Bugfix: wrong ptr to &kallocs_tree_root in task_temp_kernel_free()
task_temp_kernel_free() called bintree_find_ptr() by passing as 1st
argument &curr->kallocs_tree_root, which has type struct kernel_alloc**,
in other words, a _reference_ to a node pointer, instead of a plain node
pointer (struct kernel_alloc *), curr->kallocs_tree_root. That made the
lookup to fail when the fields around kallocs_tree_root in struct task
are re-arranged.

That's a pretty nasty bug, but the curious thing is how it even worked.
Well, kallocs_tree_root was followed by 3 bool flags: vfork_stopped,
traced and timer_ready. In 99% of the cases, all of those flags are set
to false. Also, even if rarely one of them is set (e.g. timer_ready)
we couldn't hit any problem until the "temp allocs" are used. That happens
only when select() or poll() are used. That means we needed a very unlikely
coincidence to happen in order to hit the bug. So, how that worked in the
rare cases when "temp allocs" were used, given the 3 flags as unset? Well,
in struct kernel_alloc, the bintree_node is the 1st field. Inside struct
bintree_node we have "left", "right" and "height". Passing to the function
bintree_find_ptr() a double pointer meant considering kallocs_tree_root as
the left pointer and NULL (the 3 bools + alignment) as the right pointer.
The find function performed just one more step than it could and skipped
the right (null) node. The height field it's unused in bintree_find_ptr().
While in the case of insert and remove, we passed *rightfully* the double
pointer, so in no case we had problems with the height field.