[signal] Convert ti->pending_signal to pending_signums[]
Now that we're actually going in the direction to add some basic support
custom signal handlers, it's not possible anymore to "get away" with a
simple per-task pending signal number variable. We now need to support
a bitset of signals that can be raised. While a signal handler is running,
we won't allow other signal handlers to interrupt it, but we must allow
those pending signals to get enqueued somehow for later processing.