[config][paging] Introduce KRN_PAGE_FAULT_PRINTK
As for KRN_NO_SYS_WARN, we don't really need to spam the video screen
with informative printk messages like that a process has been killed
because of a page fault. We have trace_printk() for that. But still,
in some cases such messages will remain useful. Therefore, with the new
option KRN_PAGE_FAULT_PRINTK we can configure that at compile time.