[x86][syscalls] Intr. special syscalls
With the introduction of the support for custom signal handlers,
a syscall with special needs came out: sys_rt_sigreturn(). For this
syscall, we must not process signals, nor trace it, nor enable
preemption. Initially, a special ad-hoc handling for that syscall
has been used but, in the long term, that seems a bad idea. It is
likely that other syscalls with special needs will come out and
it's much cleaner to devise a generic solution for them. That's
what the current change does: it introduces a flags field next to
each syscall function pointer, with 3 available flags:

	- SYSFL_NO_TRACE
	- SYSFL_NO_SIG
	- SYSFL_NO_PREEMPT

Most of the syscalls won't have any flags. To optimize for this case,
in handle_syscall() we check for flags == 0. In the default case, we
just call do_syscall(), otherwise we call do_special_syscall() which
will take care of the flags.