[process] Support automatic reaping
When a child process dies, its state becomes ZOMBIE and its struct process
is kept alive until some other process reads its exit code. Typically that's
supposed to be done by its parent using waitpid() but, if it doesn't do that,
that zombie process continues to exist until its parent dies and the orphaned
zombie is transferred to its closest reaper. Usually, that's PID 1, init.
At this point, PID 1 will be very likely woken up because it's supposed to
waitpid(-1, ...). [If it doesn't, the child will remain in the ZOMBIE state.]

The problem is that, sometimes, a parent is not interested at all in getting
child's exit status and it's wasteful to call waitpid() for that or, even
worse, keep all the zombies until they're passed to init. The simplest
solution for that is explicitly set SIGCHLD to SIG_IGN in the parent.

Now, when a child dies, if its parent explicitly ignored SIGCHLD, the kernel
will automatically drop the struct process. If the parent dies before the
child, the child process will still be reaped automatically if SIGCHLD was
ignored by the parent before its death.