[nanosleep] Set the `rem` param when != NULL
If nanosleep() is interrupted by a signal and the `rem` parameter is not
NULL, the kernel is supposed to set there the remaining time to complete
the sleep so that user space can restart the syscall in case the errno was
set to to EINTR. Until now, that was in the TODO list. This patch implements
this behavior.