[toolchain] Fix the build on i686 host for several apps
Tilck's build system is designed to use cross-compilers by default and
the only *officially* supported HOST_ARCH is x86_64.

Still, sometimes it's nice to be able to build Tilck on some older
i686 Linux machines. That's why Tilck supports building on i686 using
the system compiler + wrapper scripts in a unofficial, best-effort way.
This means that such scenario is not tested in CI runs and is rarely
manually tested. The likelihood of some app to not build or not work on
Tilck[1] in this case is significantly higher than in any other case.

Still, from time to time, some minor effort is spent to make things
work in this case as well. This patch, fixed the build on i686 hosts
for the following toolchain apps:

	- vim
	- lua
	- tree_cmd

[1] Why supporting i686 hosts is tricky?

The pre-compiled toolchain used by Tilck require HOST_ARCH=x86_64, so
they cannot be used. Using system's i686 compiler is not hard for
building the Tilck kernel per se, but things get tricky for user
applications, for a simple reason: Tilck supports only libmusl, while
the system compiler will very likely use glibc. Now, changing the libc
used for a GCC system compiler is definitely possible, but it requires
tricky GCC scripts to be passed to GCC and wrapper scripts such as
i686-gcc. So, in order to build an app, the build_toolchain script will
have to replace CC, with a wrapper script, ultimately calling the same
system CC, but with special options: depending on the app's build system,
that might be very tricky to make it work correctly or not a problem at all.