[toolchain] Support custom make -j <N> level for userapps with $BUILD_PAR
While it makes sense to use `make -j` for building the toolchain apps,
on some older machines with very little RAM that might not be a good
idea and cause intensive swapping. With this patch, the default behavior
won't change and `make -j` will continue to be used for all the apps and
libraries that support it but, it will be possible to customize the level
of parallelization by simply setting the BUILD_PAR environment variable.
For example, with BUILD_PAR=1, userapps will be built with `make -j1`
which is equivalent to the classic non-parallel build with `make`.