[x86/misc] Don't disable the preemption in poweroff() or reboot()
On some machines, acpi_poweroff() might actually sleep for some time
or sleep forever in case of an ACPI error and that's not possible with
preemption disabled: we end up hitting an ASSERT. With this patch,
during both poweroff() and reboot() the kernel will continue to be
fully functional for ACPI's needs. That's safe because all user tasks
are stopped, only kernel threads can run.