[timer] Fix race condition in kernel_sleep()
If the timer interrupt occurs exactly between:

    task_set_wakeup_timer(get_curr_task(), ticks);

and

    task_change_state(get_curr_task(), TASK_STATE_SLEEPING);

ti->timer_ready won't help, because in tick_all_timers() the task
still is in the RUNNABLE state. After that, it will be set to SLEEPING
and won't be woken-up by the scheduler, because it's not in the runnable
tasks list.

The fix is do exactly what is done by locking primitives:

    1. Disable the preemption
    2. Change task's state to SLEEPING
    3. Set the wakeup timer
    4. Yield while keeping the preemption disabled