[vfs] Rename struct fs to struct mnt_fs
In the long term, as the project gets contributors, it might be confusing
to use the name 'fs' to represent a *mounted* filesystem. In particular,
that could be the case once a 'struct fs_type' is introduced. Therefore,
this patch renames 'struct fs' to 'struct mnt_fs', which stands for
'mounted filesystem'.