[panic] Force enable preemption in the panic state
During panic, disable preemption count is not well-defined. For example
we might have just hit an ASSERT checking that. Therefore, we need to
set it to a sane state in order to allow the rest of the panic code
to complete, without hitting other ASSERTs (or, at least, reduce the
likelihood of that to happen).