[tests] Make the fork_oom test to estimate the amount of memory
Previously, the test simply assumed that we a had a bit less than 128 MB
of total memory.  That was fine when QEMU was run with the default
options. When the test is run on a physical machine or on a VM with
more memory, the test didn't work properly simply because it could
commit more than the hard-coded 96 * 2 * MB value.