[config] Intr. FB_CONSOLE_FAILSAFE_OPT
On modern machines, the loop unrolling in fb_draw_char_failsafe()
improves the performance of the fb_console by 0.3%, while it adds
+3.5KB to the binary and it's used only in the failsafe case.

It's reasonable to have a more compact failsafe function by default.
On older machines, the loop unrolling optimization gives about 6.5%
of speedup; still not much, but on some even older machines, the
impact might be even bigger. Therefore, it's worth keeping the
optimization in the code under a compile-time option.