[paging] Make pageframes_refcount use 32-bit integers
Even if most of the page frames won't have a high ref-count, there is
one particular physical page which it's likely to be referenced a lot:
the zero_page. Actually, if a machine has more than 256 MB of memory and
user space apps use a lot of over-commited memory, the 16-bit unsigned
ref-count will wrap-around with CATASTROPHIC effects; for example, when
it gets to 1, handle_potential_cow() will allow user apps to directly take
ownership of the zero_page.