[tracing] Fix save_param_buffer() in the case a C string buf faults
The function correctly handled the case where reading a sized-buffer
parameter causes a PF or attempts to read unaccessible memory, but
it missed the case where the buffer is a (NUL-terminated) C string and
tried to use strlen() on it. The fix was simply to use the function
copy_str_from_user(), designed for this purpose.