[cmake] Fix build failure with USE_SYSCC=1 and GCC 9.4.0
Using system's compiler to build Tilck is a special dev-feature and
it is not _officially_ supported. Because of that, sometimes, despite
being tested on CI, it still breaks on some slightly-different build
configurations. This patch fixes a failure of this type by disabling
_FORTIFY_SOURCE, because it requires extra symbols like __fdelt_chk
not availabile in the kernel, nor in libgcc.