[cmdline] Re-design the support for kernel cmdline options
Previously, each kernel option required manually-written dedicated code.
This was good enough for just a few options, but with time Tilck requires
more and more options and the current implementation simply doesn't scale.

This patch re-designs the support for cmdline options, by allowing them
to be defined in a "codeless" generic way, using some macro tricks.
The only special option remanins "-cmd" (with the new alias "--") which
can be used only as last.