[shutdown] Make stop_all_user_tasks() to skip the current task
If the current task is stopped, it could be preempted and than
never get to complete the shutdown process. Typically, it's the
init process (pid 1) which calls the reboot() syscall and we
already checked for that in stop_all_user_tasks() but, in case
the user runs `reboot -f` or `poweroff -f`, the signal to `init`
is bypassed and the current task becomes the one that must "live
until the end".