[runners] Re-run once more tests failing with shell_unknown_exit_code
From time to time, the per-second message by the "sched alive thread"
interleaves the shell's exit code like that:

	[init] the shell with [    2.139] ---- Sched alive thread: 2 ----
	pid 2 exited with status: 0

And this prevents the runner to correctly parse shell's exit code,
resulting in a false positive failure.

To mitigate this problem, let's re-run the test just once, if it has
failed with: shell_unknown_exit_code.