[docs] Replace `make -j` with `make` in the docs (issue #101)
Even if parallel builds of Tilck work perfectly, on machines with a
scarse amount of RAM (e.g. < 4 GB), the maxiumum level of parallelism
(make -j) might exahust all available memory and trigger the Linux
OOM killer. Technically, this has nothing to do with Tilck, and it
happens with *any* other big-enough software project, but still it's
better not to suggest everyone to use `make -j` in the documentation
files. Experienced developers are already aware of such things and will
use the right level of parallelism for their machine.