[devshell] De-dup the list of tests by using a separate header file
Currently, in order to register a new system test as shellcmd it is
necessary to both declare its entry point with DECL_CMD() and then
to list it in the `_cmds_table` array with CMD_ENTRY(). As the list
gets longer and longer this duplication looks uglier and uglier.
This change moves the whole list to a dedicated header file called
cmds_table.h that will be included twice by cmds_table.c: once for the
declarations, once for the entries in the `_cmds_table`.