[process32] Make switch_to_task() to fully support ti == curr
In the corner case where there are pending signals and the current process
will be re-scheduled to run, do_schedule() needs to call switch_to_task()
passing the current task, because the task has to be setup for signal
processing and cannot just return as if nothing happened. In order to
support that, switch_to_task() needed a few minor changes.