[64-bit] Make Tilck to simply compile for x86_64
With the current patch, the Tilck kernel and all the additional toolchain
packages can compile for x86_64, on x86_64 systems (our BUILD_ARCH),
using the standard cross-compilers from toolchains.bootlin.com.

Of course, nothing really works (yet). The kernel needs a lot of additional
code in order to work. For the moment, stub functions have been added just
to make it compile.

NOTE: this patch also modifies a bit the format of the toolchain, by renaming
$TCROOT/.gcc_tc_ver to $TCROOT/.gcc_tc_ver_$ARCH: this change is essential
in order to allow the same toolchain to be used for different target
architectures. The broken assumption made in the past was that once a
GCC_TC_VER is picked, it could be used for all the architectures, while the
pre-built toolchains we use have actually separate lists of compilers for each
target architecture: there's *no guarantee* that the same compiler version is
available for all the achitectures we need.