[interactive] Re-try once in case of test failure
While running in CI like Azure Pipelines, timing is uncertain:
because of that, it happens (rarely!) that some of the most
demaning interactive tests like `vim` fail because the output
doesn't match exactly the expected one.

A workaround to this problem could be to increase the internal
delays, making everything to go slower: that might accomodate
the CI case in the < 1% of the cases, but it will make the test
run significantly slower in 99% of the cases. Therefore, we're
taking the opposite approach: run with the same speed and take 2x
of the time, in case of failure.