[gtests] Intr. kernel test headers
Currently, static kernel functions can be called by unit tests simply by
making them STATIC, which evaluates to an empty token in the context of
unit tests, making the symbol visible. The problem is that we still
needed to declare those (private) functions in the unit tests, in order
to call them and that's unsafe because a change in the function signature
will not result in any compile-time error or warning. To avoid this kind
potential issues, we're introducing a set of "test" kernel headers: those
headers will contain the signatures of the STATIC functions and will be
included both by the kernel code and the unit tests. This way, it won't
be possible anymore to have a mismatch in the function signature without
getting a compile error.