[64-bit] Make kernel_va != KERNEL_BASE_VA
On i386, KERNEL_BASE_VA (+3 GB) is used both as a base VA for the linear
mapping of the physical memory and for placing the .text and .data of the
kernel itself. In other words, the kernel itself is mapped only *once* in
the linear virtual address space starting at KERNEL_BASE_VA, at the offset
KERNEL_PADDR (+1 MB). Therefore, we don't need a dedicated mapping for the
kernel itself.

On x86-64, not only we'd like to have KERNEL_BASE_VA much bigger to
order to allow user apps to have much larger virtual address space,
but even if we kept KERNEL_BASE_VA at +3 GB we couldn't link with the
default memory model (-mcmodel=small), simply because it requires both
the text and data to be within the lower 2 GB of the address space.
Using -mcmodel=large immediately solves the problem, but comes at a high
runtime overhead.

Therefore, the solution is simply to separate the VA of kernel's text
and data from the starting of the linear mapping. This way, the kernel
will be mapped twice: once at -2 GB (0xFFFFFFFF80000000) and the other
at KERNEL_BASE_VA + KERNEL_PADDR. This approach is very flexible as it
will allow the kernel to be placed at runtime at any PADDR, as the
virtual mapping at -2 GB does not depend on any hard-coded PADDR, so
the generated code will not be affected.