[cmake] Patch with TOOL_WS each module linked to gtests
When the symbol weakening was introduced in Tilck to allow wrapping of
symbols in the same translation unit, the TOOL_WS (weaken_syms) was
run only against the main `kernel_for_test` static library. Now that
the technology proved to be useful in practice with GMock, it's time
to patch also the _noarch static libraries generated by modules that
are linked to the gtest executable. This will allow unit tests to
mock symbols in all the whitelisted "noarch" modules. Currently they
are:

  - console
  - tracing
  - sysfs

Note: this change moves the WRAPPED_SYMS list to a dedicated file:
other/cmake/wrapped_syms.cmake.