[toolchain] Check for spaces in the $MAIN_DIR and abort
Tilck's build_toolchain script and build system doesn't support paths
with spaces as doing so would require an insane amount of ugly quoting
and not only that would not be worth the effort, but the final code would
look would even worse than now. Therefore, for those reasons, compiling
this project will have the reasonable requirement of checking out the
source in a path free of spaces.

This simple change adds a check for that corner case and fails
gracefully with a human-friendly error in case the project has been
checked out in a path with spaces.