[uefi] Add support for Runtime Services in the kernel
Multiboot 1.0 protocol doesn't support extra fields.
So, new struct tilck_extra_boot_info was introduced
and set to `apm_table`.

When SetVirtualAddressMap() is called, it requires the
physical addresses to be directly accessible. So, they
were identity mapped before the call and removed
after the call.

The read-only EFI Runtime memory regions were
considered as Code regions and the rest as Data
regions.

This doesn't work for 64-bit yet. That will be
implemented when Tilck is ported to 64-bit.