[datetime] Drop the `weekday` from the struct datetime
Since the `weekday` field is not returned by the UEFI GetTime() RT
function, in order to maintain it the struct datetime we'd need
to perform a relatively expensive computation. It seems like we
don't really need this field in our datetime struct. Also, given
such a datetime it's always possible to compute the `weekday` if
that is really necessary.