[acpi] Allow building ACPICA with -Wno-error
Tilck's ACPICA is from 2020 and triggers warnings on new compilers. Because
of -Werror, this leads to a build failure. Until ACPICA's source is upgraded,
it makes sense to at least allow the build to succeed, keeping the warnings
visible.

BONUS: "g++-multilib" is mentioned in the list of packages necessary for
building all the non-default Tilck configurations, including the ones
with system's clang compiler etc.