[64-bit] Use separate vaddrs for reading phy. mem. and for the kernel
BASE_VA is used for reading the physical memory and
KERNEL_VADDR is used for the kernel.

In 64-bit, we want to map the kernel at -2GB to be
able to use 32-bit offsets. Hence, a separate vaddr
was introduced for achieving this.

To avoid 64-bit specific tricks, this change is
generic and for both 32-bit and 64-bit kernels.