[64-bit] Allow Tilck to be mapped at a non-linear vaddr
This is the part of two of commit 29ee48c7f3. Here we completely decouple
the default linear mapping of the physical address space from the kernel's
vaddr of text, data & bss. By default, the 32-bit Tilck kernel will continue
to be mapped at BASE_VA + KERNEL_PADDR, using the general-purpose linear
mapping, but when the KRN32_LIN_VADDR option is disabled, the kernel itself
will be mapped at a much higher address, near LINEAR_MAPPING_END.

The reason is the same as 29ee48c7f3: the 64-bit kernel will need to be mapped
well outside of the 4 GB limit in order to support more physical memory in the
linear mapping. But, doing that while keeping the linear mapping will be
extremely inefficient because all the major compilers will refuse to compile
an ELF binary placed at such a high vaddr without using -mcmodel=large. And
the problem with that is that -mcmodel=large uses inefficient instructions to
access global variables and call functions because it doesn't assume that data
and text are within 2 GB of any point. It assumes that each variable or function
can be *anywhere* in the 64-bit virtual address space, while that is not true.
In theory, it's totally possible to patch the compiler and make it use a small
memory model at a high address, but why bothering with that? The mainstream
solution is simply to place the kernel somewhere within [-2 GB, 0). This allows
both using the smaller (and more efficient) memory model and to leave all the
virtual space we want for user programs.

This change makes mapping the kernel at a non-linear vaddr possible, by removing
all the hard-coded assumptions about kernel's base vaddr being within the
generic linear mapping at BASE_VA. The core idea is to rename KERNEL_BASE_VA
to BASE_VA by fixing most of the instances to use the new BASE_VA and
selectively use the new KERNEL_BASE_VA/KERNEL_VADDR when necessary. One kind of
cases for the new KERNEL_BASE_VA is when we're accessing kernel's symbol table:
for that we really do care about kernel's vaddr. Another case is when we use
the address of kernel buffers (e.g. kmalloc's first heap, zero_page, kpdir_buf)
in contexts where later a VA -> PA conversion will occurr because the code works
mostly with heap-allocated buffers, which are mapped using the linear mapping.
In such cases, since the VA does not start at BASE_VA but at KERNEL_BASE_VA,
the conversion leads to a wrong PA and that triggers any sort of failures.
The correct approach for those special cases is to convert the kernel VA to a
linear VA with a double-conversion: kernel symbol VA -> PA -> linear VA.
Therefore, we'll access our own kernel variables through a different mapping
because it allows the rest of the generic code to work. That requires also
keeping a mapping of the kernel in the early paging setup in start.S, in
addition to the dedicated kernel mapping near LINEAR_MAPPING_END.