[toolchain] Use the SYSCC only when $USE_SYSCC == 1, no matter the $ARCH
Currently, in many places we check for $HOST_ARCH == $ARCH and use the system
compiler in that case, because that check has been written for the
$HOST_ARCH == i386 case, given that the target $ARCH could only possibly be i386.
Now that we're working on ARCH=x86_64, we definitively don't want the system
compiler to be the automatically the default case because it's always a bit
tricky to support. Tilck has one officially supported way to be built and that's
by downloading the pre-compiled toolchain, no matter the target architecture.
The support for a build with a system compiler is limited and doesn't need to
work in all cases (all the toolchain packages).