[64-bit] Simplify start.S and use EDX instead of EDI
Currently the x86-64 start.S file uses the EDI register for various operations,
while that is not necessary. For example, there is no need to zero the buffers
used for the page tables, simply because they are in the .bss section and are
zeroed by the bootloader. After removing that code, all uses of EDI can be
replaced with EDX, preserving the value of EDI. That register will be used in
the next commit to store the value of the first parameter of kmain(), following
the System V ABI.