[64-bit] Make start.S to use dedicated buffers for the early page tables
Currently, the x86-64 start.S uses 12 KB out of `page_size_buf` which is just
4 KB big. That's a bug introduced with commit 165e54e449b. That causes an
overflow which has weird side-effects when the multiboot info is used and
variables immediately after `page_size_buf` are touched, simply because doing
so corrupts the early page tables.

The fix is to use two dedicated page-size buffers, only for the x86-64 case.
In addition to fixing that, this change fixes start.S to correctly set the
first two arguments of kmain() in the RDI and RSI registers so that the
multiboot info can be read.

Now when the kernel panics due to NOT_IMPLEMENTED, we can see also the system's
memory map, provided to us by the bootloader.