[64-bit] Make the bootloader to offer only text-mode for x86-64
We don't support the framebuffer console on x86-64 simply because we don't
have any form of paging implementation that we need absolutely need to map
the framebuffer at a convenient address in kernel's virtual address space.
Therefore, since that we must use text mode for the moment, just make it the
default offered by the bootloader, so that we won't have to select it manually
every time.