[gcov] Prefix the gcov cmds with "tilck_"
The Tilck cmds defined in tilck_cmd.c are not real syscalls. They are cmds
under a single syscall, sys_tilck_cmd, which is number 499 on i386 at the
moment. Because of that, they MUST not use the "sys_" prefix, as it easier
for multiple purposes assume that all the syscalls and nothing else in the
kernel start with "sys_".