[tracing] Replace the `tracing_lock` with IRQ disabling
This change allows tracing events to be registered even by IRQ handlers.
While that seemed unnecessary for a long time, it turned out that's necessary
in order to make trace_printk() useable from IRQ handlers. Why that is useful?
Because we would like to make the printk() messages be also saved in the trace
buffer, at least in some debug configurations. The plan to make that possible
is to use trace_printk(), but that need to work everywhere where printk() can
and that includes IRQ handlers.