[toolchain] Fix support for USE_SYSCC=1 on HOST_ARCH=x86_64
This change fixes the broken support for USE_SYSCC=1. This rarely-used option
gives us two features:

   - building the kernel with the system compiler, which can also be clang
     and benefit from extra checks & warnings

   - building the user apps with the our custom libmusl instead of one the
     one included in the pre-built toolchain. The advantage of that is having
     the debug symbols in case of low-level libc issues to debug.

Building the user apps entirely with the system compiler is not part of the
goals of this feature and not only isn't very useful as a feature unless
$ARCH == $HOST_ARCH, but it's extremely hard to make it work, because of the
libmusl. Cross-compiling to from x86-64 to i386 when gcc-multlib is installed
is simple with the -m32 option, but doing so while also changing the libc
to libmusl opens several nastly problems, because when we pass the libmusl
specs to the system compiler, somehow at least for BusyBox the 64-bit system
linker (ld) ends up being called, instead of the 32-bit system linker, which
typically is /lib32/ld-linux.so.2.

Therefore, since we're supporting USE_SYSCC=1 only for the two features above,
we solve this problem with an ugly hack: use the GCC cross-compiler from the
toolchain, which by default links to the target ARCH and change the libc
using the musl specs from our custom libmusl build.

Over the time, this feature has been broken for multiple reasons, when changes
have been made to fix one use-case (e.g. ARCH == HOST_ARCH or ARCH == x86_64
or HOST_ARCH == aarch64) forgetting about the undesired side-effects for the
rest of the use cases.

Recently, commit 37a1b27a7 completely broke the build of USE_SYSCC=1 in the
most important use-cases, when HOST_ARCH == x86_64 and ARCH == i386. This
change hopefully fixes most of not all of the issues related with this feature
in one big change.

In addition to that, this change makes the diagnosis of ugly build failures of
user apps easier, by writing to the build.log file in each package a fair
amount of debug data related to the environment variables.